refactor: store 파일명 수정 deviceDetail -> deviceTabs
